Display bug state errors with alert dialog
For example drag and dropping an issue from some column states to another
(i.e. from open all the way to closed) is not allowed by the bugzilla
lifecycle rules and bz will return an error when you try doing so.
We used to display this error message but the change for #80 introduced a
regression.
With this change we can now see the error messages again by using a callback.